requirements.txt
flask==3.0.3
flask-cors==4.0.1
PyPDF2==3.0.1
python-docx==1.1.2
scikit-learn==1.5.1
numpy==1.26.4
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.0
